Quest finished with the option

+++ upload.php
@@ -4,6 +4,12 @@
 // liste image upload
 $file = new FilesystemIterator('uploads');
 
+if (isset($_POST['submit'])) {
+    if(unlink($_POST['submit'])) {
+        header("Location:upload.php");
+    }
+}
+
 // chemin vers un dossier sur le serveur qui va recevoir les fichiers uploadés
 // (attention ce dossier doit être accessible en écriture)
 $uploadDir = 'uploads/';
@@ -25,12 +31,12 @@ for($i=0; $i < count($_FILES['avatar']['name']); $i++) {
     
     // Si l'extension n'est pas dans le tableau
     if(!in_array($extension, $extensions)) { 
-        $erreur =  'Votre fichier ' . $filename[$i] . ' n\'est pas conforme' .'<br>' .
+        $error =  'Votre fichier "' . $filename[$i] . '" n\'est pas conforme' .'<br>' .
         'Vous devez uploader un fichier de type png, gif, jpg, txt ou doc...';
     }
     
     if($taille > $taille_maxi) {
-        $erreur = 'Votre fichier ' . $filename[$i] . ' n\'est pas conforme' .'<br>' .
+        $error = 'Votre fichier "' . $filename[$i] . '" n\'est pas conforme' .'<br>' .
         'Le fichier est trop gros...';
     }
     
@@ -42,17 +48,20 @@ for($i=0; $i < count($_FILES['avatar']['name']); $i++) {
         
         // Si la fonction renvoie TRUE, c'est que ça a fonctionné...
         if(move_uploaded_file($_FILES['avatar']['tmp_name'][$i], $uploadDir . $filenameUniq)) {
-            echo 'Upload du fichier ' . $filenameRetour . ' effectué avec succès !' .'<br>';
+            echo 'Upload du fichier "' . $filename[$i] . '" effectué avec succès !' .'<br>';
         
         // Sinon (la fonction renvoie FALSE).
         } else {
             
-            echo 'Echec de l\'upload !';
+            echo 'Echec de l\'upload "' . $filename[$i] . '" !'  .'<br>';
         }
     } else {
-        echo $erreur;
+        echo $error;
     }
 }
+
+
+
 
 ?>
 
@@ -73,8 +82,8 @@ for($i=0; $i < count($_FILES['avatar']['name']); $i++) {
             <!-- On limite le fichier à 100Ko -->
             <h2 style="text-align:center; color:yellow" for="imageUpload">Upload image</h2></br>
             <input type="hidden" name="MAX_FILE_SIZE" value="1000000" />
-            Fichier : <input type="file" name="avatar[]" multiple="multiple" /></br>
-            <input type='submit' name='submit' value='Upload'>
+            <input class="btn btn-light" type="file" name="avatar[]" multiple="multiple" /></br>
+            <button type="submit" class="btn btn-success">Upload</button>
             <div>
                 <a style="color:red" href="upload.php">Rafraichir</a>
             </div>
@@ -84,14 +93,16 @@ for($i=0; $i < count($_FILES['avatar']['name']); $i++) {
         <br><h2 style="text-align:center; color:yellow">Avatar en stock :</h2>
     </div>
     <div style="">
-        <form method="post" action="upload.php">
+        <form  style="" method="post" action="upload.php">
         <?php
         foreach ($file as $fileinfo) {
             echo '<figure class="col-3" style="border:thin #c0c0c0 solid; display:inline-flex; flex-flow:column; padding: 5px; max-width: 220px; margin: auto;">';
             echo '<img style="max-width:220px; max-height:150px;" src="uploads/' . $fileinfo->getFilename() . '" />';
             echo '<figcaption style="background-color:#222; color:#fff; font:italic smaller sans-serif; padding:3px; text-align:center;">' . $fileinfo->getFilename() . '</figcaption>';
-            echo '<input class="btn btn-danger" type="submit" name="submit" value="Delete" />Delete';
-            echo '</figure>';    
+            if (file_exists($fileinfo)) {
+                echo '<button class="btn btn-warning" type="submit"  name="submit" value="' . $fileinfo . '">Delete</button>';
+                echo '</figure>';
+            } 
         }
         ?>
         </form>
